Rename to index.js
